requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.3
colorama==0.4.6
cryptography==47.0.0
ecdsa==0.19.2
fastapi==0.136.1
flexcache==0.3
flexparser==0.4
greenlet==3.5.0
h11==0.16.0
httptools==0.7.1
idna==3.13
obd==0.7.3
passlib==1.7.4
Pint==0.24.4
platformdirs==4.9.6
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyserial==3.5
python-dotenv==1.2.2
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
watchfiles==1.1.1
websockets==16.0
